# docs/swagger.yaml
definitions:
...
  common.ServerErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.UserInfo:
    properties:
      address:
        $ref: '#/definitions/gocloak.UserInfoAddress'
      email:
        type: string
      email_verified:
        type: boolean
      family_name:
        type: string
      gender:
        type: string
      given_name:
        type: string
      locale:
        type: string
      middle_name:
        type: string
      name:
        type: string
      nickname:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      picture:
        type: string
      preferred_username:
        type: string
      profile:
        type: string
      sub:
        type: string
      updated_at:
        type: integer
      website:
        type: string
      zoneinfo:
        type: string
    type: object
...
  credential.CredentialConfiguration:
    properties:
      claims:
        additionalProperties: true
        type: object
      credential_definition:
        $ref: '#/definitions/credential.CredentialDefinition'
      credential_signing_alg_values_supported:
        items:
          type: string
        type: array
      cryptographic_binding_methods_supported:
        items:
          type: string
        type: array
      display:
        items:
          $ref: '#/definitions/credential.LocalizedCredential'
        type: array
      format:
        type: string
      order:
        items:
          type: string
        type: array
      proof_types_supported:
        additionalProperties:
          $ref: '#/definitions/credential.ProofType'
        type: object
      schema:
        additionalProperties: true
        description: /Out of OID Spec, but useful
        type: object
      scope:
        type: string
      topic:
        description: Subject of the credential within the system
        type: string
      vct:
        type: string
    type: object
  credential.CredentialDefinition:
    properties:
      credentialSubject:
        additionalProperties:
          $ref: '#/definitions/credential.CredentialSubject'
        type: object
      type:
        items:
          type: string
        type: array
    type: object
  credential.CredentialOffer:
    properties:
      credential_offer:
        type: string
      credential_offer_uri:
        type: string
    type: object
...
  credential.CredentialSubject:
    properties:
      display:
        items:
          $ref: '#/definitions/credential.Display'
        type: array
    type: object
  credential.DescriptiveURL:
    properties:
      alternative_text:
        type: string
      url:
        type: string
    type: object
  credential.Display:
    properties:
      locale:
        type: string
      name:
        type: string
    type: object
...
  credential.LocalizedCredential:
    properties:
      background_color:
        type: string
      locale:
        type: string
      logo:
        $ref: '#/definitions/credential.DescriptiveURL'
      name:
        type: string
      text_color:
        type: string
    type: object
...
  credential.ProofType:
    properties:
      proof_signing_alg_values_supported:
        items:
          type: string
        type: array
    type: object
...
  gocloak.UserInfoAddress:
    properties:
      country:
        type: string
      formatted:
        type: string
      locality:
        type: string
      postal_code:
        type: string
      region:
        type: string
      street_address:
        type: string
    type: object
...
  handlers.IssueCredentialRequestBody:
    properties:
      payload:
        additionalProperties: true
        type: object
      type:
        type: string
    type: object
  handlers.ListCredentialRequestBody:
    properties:
      search:
        type: string
    type: object
...
  handlers.SaveConfigrationsRequest:
    properties:
      historyLimit:
        type: integer
      language:
        type: string
    required:
    - historyLimit
    - language
    type: object
...
  presentation.Description:
    properties:
      format:
        type: string
      id:
        type: string
      name:
        type: string
      purpose:
        type: string
    type: object
...
  presentation.FilterResult:
    properties:
      credentials:
        additionalProperties: true
        type: object
      description:
        $ref: '#/definitions/presentation.Description'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the personal credential manager cloud services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Account service API
  version: "1.0"
paths:
  /configurations/getUserInfo:
    get:
      consumes:
      - application/json
      description: Retrieves user information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get user information
      tags:
      - configurations
  /configurations/list:
    get:
      consumes:
      - application/json
      description: Retrieves configurations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get configurations
      tags:
      - configurations
  /configurations/save:
    post:
      consumes:
      - application/json
      description: Saves configurations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Configurations
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.SaveConfigrationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Save configurations
      tags:
      - configurations
  /credentials/issue:
    post:
      consumes:
      - application/json
      description: Requests the issuance of a credential
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: type and credential subject
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.IssueCredentialRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credential.CredentialOffer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Request issuance of a credential
      tags:
      - credentials
  /credentials/list:
    get:
      consumes:
      - application/json
      description: Lists all credentials
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Search
        in: body
        name: search
        schema:
          $ref: '#/definitions/handlers.ListCredentialRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.FilterResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: List credentials
      tags:
      - credentials
...
  /credentials/schemas:
    get:
      consumes:
      - application/json
      description: Retrieves all credential configurations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/credential.CredentialConfiguration'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get credential configurations
      tags:
      - credentials
...
  /presentations/list:
    get:
      consumes:
      - application/json
      description: Lists all presentations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Search
        in: body
        name: search
        schema:
          $ref: '#/definitions/handlers.ListCredentialRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.FilterResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: List presentations
      tags:
      - presentations
...
swagger: "2.0"